Minor comments, added readme, gitignore and outputs
