+++ output/code.cs
@@ -1,32 +1,6 @@
 using System.Collections.Generic;
 using System;
 
-public class Author {
-    public string Name { get; set; }
-    public string Email { get; set; }
-    public List<Book> Books { get; set; }
-
-    public Author(string name, string email)
-    {
-        this.Name = name;
-        this.Email = email;
-        this.Books = new List<Book>();
-    }
-
-    public void AddBook(Book book)
-    {
-        if (!Books.Contains(book))
-        {
-            Books.Add(book);
-            book.AddAuthor(this);
-        }
-    }
-
-    public override string ToString()
-    {
-        return "Author{Name=" + Name + ", Email=" + Email + ", Books=" + Books.Count + "}";
-    }
-}
 public class Library {
     public string Name { get; set; }
     public string Address { get; set; }
@@ -54,19 +28,19 @@ public class Library {
     }
 }
 public class Book {
-    public int Pages { get; set; }
-    public string Title { get; set; }
     public DateTime Release { get; set; }
-    public List<Author> Authors { get; set; }
+    public string Title { get; set; }
+    public int Pages { get; set; }
     public Library Library { get; set; }
+    public List<Author> Authors { get; set; }
 
-    public Book(int pages, string title, DateTime release)
+    public Book(DateTime release, string title, int pages)
     {
-        this.Pages = pages;
-        this.Title = title;
         this.Release = release;
-        this.Authors = new List<Author>();
+        this.Title = title;
+        this.Pages = pages;
         this.Library = null;
+        this.Authors = new List<Author>();
     }
 
     public void AddAuthor(Author author)
@@ -80,6 +54,32 @@ public class Book {
 
     public override string ToString()
     {
-        return "Book{Pages=" + Pages + ", Title=" + Title + ", Release=" + Release + ", Authors=" + Authors.Count +  ", Library=" + (Library != null ? Library.Name : "None") + "}";
+        return "Book{Release=" + Release + ", Title=" + Title + ", Pages=" + Pages + ", Library=" + (Library != null ? Library.Name : "None") +  ", Authors=" + Authors.Count + "}";
+    }
+}
+public class Author {
+    public string Email { get; set; }
+    public string Name { get; set; }
+    public List<Book> Books { get; set; }
+
+    public Author(string email, string name)
+    {
+        this.Email = email;
+        this.Name = name;
+        this.Books = new List<Book>();
+    }
+
+    public void AddBook(Book book)
+    {
+        if (!Books.Contains(book))
+        {
+            Books.Add(book);
+            book.AddAuthor(this);
+        }
+    }
+
+    public override string ToString()
+    {
+        return "Author{Email=" + Email + ", Name=" + Name + ", Books=" + Books.Count + "}";
     }
 }
